"naif-programmer" v0.1.1
When calling a object role set to null that share a method with the same
signature as a lower priority one not null the null was being resolved
leading to a NPE.

When a user was not declaring a rigid type for a role type and calling a
method of the rigid type from the role type a StackOverflowException was
thrown. A exception is still thrown but is more descriptive regarding the
problem.

New test cases added.