"big-inc" v0.8.0
Improve interface of the register

Add new capabilities to the register

Enhance the bypass of java access checking

New core injection method

Callback methods for rigid injection

Turn on and off the role based on a reference to rigid

It is possible to define method to trigger injection of rigid & is also
possible to select if the code is to run after or before the method. For
constructors it is not possible to define the order. The code is executed
always after.

Extend anotations to overriden methods automatically

Addes feature to call super method automatically